refactor: 커밋 테이블 row 컴포넌트 분리
